puzzle

+++ 15Puzzle/src/main/java/de/amr/games/puzzle15/PuzzleSimulator.java
@@ -4,18 +4,17 @@ import java.awt.EventQueue;
 
 import javax.swing.JFrame;
 
-
 public class PuzzleSimulator extends JFrame {
-	
+
 	public static void main(String[] args) {
 		EventQueue.invokeLater(PuzzleSimulator::new);
 	}
-	
+
 	private Puzzle puzzle;
 	private PuzzleView view;
-	
+
 	public PuzzleSimulator() {
-		puzzle = new Puzzle(4);
+		puzzle = new Puzzle((byte) 4);
 		view = new PuzzleView(puzzle, 100);
 		setTitle("15-Puzzle");
 		setDefaultCloseOperation(EXIT_ON_CLOSE);
